# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Snappfood API",
        "description": "API documentation for the Snappfood.",
        "contact": {
            "email": "farshidboroomand@gmail.com.com"
        },
        "version": "V1"
    },
    "paths": {
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a paginated list of users",
                "description": "Returns a paginated list of all registered users.",
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "2bb3ded7-00be-4763-8016-9d072ba374d3"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Charlie"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Stehr"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "clemmie.bartell@example.org"
                                                    },
                                                    "mobile": {
                                                        "type": "string",
                                                        "example": "+1-240-842-4579"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve users"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a new user with the given information.",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "mobile",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "farshid"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "boroomand"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "farshid@gmail.com"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "9124049032"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "userId": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "4ebe61f9-97de-4505-a4af-338559cce665"
                                                },
                                                "firstName": {
                                                    "type": "string",
                                                    "example": "farshid"
                                                },
                                                "lastName": {
                                                    "type": "string",
                                                    "example": "boroomand"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "farshid@gmail.com"
                                                },
                                                "mobile": {
                                                    "type": "string",
                                                    "example": "9124049032"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The first name field is required."
                                                    }
                                                },
                                                "last_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The last name field is required."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "mobile": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The mobile field is required."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}